UIImageView category to override Invert Colors.
For a variety of reasons, individuals with low vision often benefit from the "Invert Colors" mode available on all iOS devices. Toggling this setting (General > Accessibility > Invert Colors) maps white to black, blue to orange, etc. The mainstream press sometimes calls this "night mode." While the setting can improve text legibility and reading comfort, it leaves images a garbled, ghostly mess.

This category offers a way for developers to prevent content images from inverting along with the rest of the screen. This directly overrides the user's preference and may be undesired, so use it with care!

Potential enhancements:
- Dim images that have been un-inverted.
- Only toggle colors on tap or long press (thanks @chrishoog).